PLT-8077 - Reduced custom To/FromJSON instances as much as possible
