+++ marconi-chain-index/marconi-chain-index.cabal
@@ -70,6 +70,9 @@ library
     Marconi.ChainIndex.Api.JsonRpc.Endpoint.MintBurnToken
     Marconi.ChainIndex.Api.JsonRpc.Endpoint.TargetAddresses
     Marconi.ChainIndex.Api.JsonRpc.Endpoint.Utxo
+    Marconi.ChainIndex.Api.JsonRpc.Endpoint.Utxo.SpentInfoResult
+    Marconi.ChainIndex.Api.JsonRpc.Endpoint.Utxo.Types
+    Marconi.ChainIndex.Api.JsonRpc.Endpoint.Utxo.Wrappers
     Marconi.ChainIndex.Api.JsonRpc.Routes
     Marconi.ChainIndex.Api.JsonRpc.Server
     Marconi.ChainIndex.Api.Rest.Endpoint.Metrics
